Complete whole-turn mechanics, including psychic-dependant buffs and charge distances.
Also change the original shrinkModels optimization, as folding is now
much faster than casing the joint binomial distributions (dakkabots are
a clear example of this). Instead, evaluate the attack of a single
model/weapon in each group and then replicate it by the number of
models/weapons in the group.